refactor: organizes UserProfileCard exports in their respective folder
