+++ src/components/admin/user-profile-card/index.ts
@@ -1,0 +1,3 @@
+export { default as UserProfileCard } from './UserProfileCard.vue';
+export { default as UserButton } from './UserButton.vue';
+export type { UserProfileCardProps, UserButtonProps, UserCard, UserProfileCardPropsDropdownItem } from './types';
